extract shared function to compute whether next action should be force push
