+++ app/src/lib/rebase.ts
@@ -1,10 +1,14 @@
-import { IRepositoryState, RebaseConflictState } from '../lib/app-state'
+import {
+  IRepositoryState,
+  RebaseConflictState,
+  IBranchesState,
+} from '../lib/app-state'
 import {
   ChooseBranchesStep,
   RebaseStep,
   ShowConflictsStep,
 } from '../models/rebase-flow-step'
-import { Branch } from '../models/branch'
+import { Branch, IAheadBehind } from '../models/branch'
 import { TipState } from '../models/tip'
 import { clamp } from './clamp'
 
@@ -65,4 +69,31 @@ export function initializeRebaseFlowForConflictedRepository(
  */
 export function formatRebaseValue(value: number) {
   return Math.round(clamp(value, 0, 1) * 100) / 100
+}
+
+/**
+ * Check application state to see whether the action applied to the current
+ * branch should be a force push
+ */
+export function isCurrentBranchForcePush(
+  branchesState: IBranchesState,
+  aheadBehind: IAheadBehind | null
+) {
+  if (aheadBehind === null) {
+    // no tracking branch found
+    return false
+  }
+
+  const { tip, rebasedBranches } = branchesState
+  const { ahead, behind } = aheadBehind
+
+  let branchWasRebased = false
+  if (tip.kind === TipState.Valid) {
+    const localBranchName = tip.branch.nameWithoutRemote
+    const { sha } = tip.branch.tip
+    const foundEntry = rebasedBranches.get(localBranchName)
+    branchWasRebased = foundEntry === sha
+  }
+
+  return branchWasRebased && behind > 0 && ahead > 0
 }
