+++ app/src/ui/app.tsx
@@ -98,6 +98,7 @@ import { RebaseFlow, ConfirmForcePush } from './rebase'
 import {
   initializeNewRebaseFlow,
   initializeRebaseFlowForConflictedRepository,
+  isCurrentBranchForcePush,
 } from '../lib/rebase'
 import { BannerType } from '../models/banner'
 
@@ -1944,18 +1945,14 @@ export class App extends React.Component<IAppProps, IAppState> {
     const rebaseInProgress =
       conflictState !== null && conflictState.kind === 'rebase'
 
-    const { pullWithRebase, tip, rebasedBranches } = state.branchesState
+    const { aheadBehind, branchesState } = state
+    const { pullWithRebase, tip } = branchesState
 
     if (tip.kind === TipState.Valid && tip.branch.remote !== null) {
       remoteName = tip.branch.remote
     }
-    let branchWasRebased = false
-    if (tip.kind === TipState.Valid) {
-      const localBranchName = tip.branch.nameWithoutRemote
-      const { sha } = tip.branch.tip
-      const foundEntry = rebasedBranches.get(localBranchName)
-      branchWasRebased = foundEntry === sha
-    }
+
+    const isForcePush = isCurrentBranchForcePush(branchesState, aheadBehind)
 
     return (
       <PushPullButton
@@ -1969,7 +1966,7 @@ export class App extends React.Component<IAppProps, IAppState> {
         tipState={tip.kind}
         pullWithRebase={pullWithRebase}
         rebaseInProgress={rebaseInProgress}
-        branchWasRebased={branchWasRebased}
+        isForcePush={isForcePush}
       />
     )
   }
